serialize

+++ Source/Engine/Container/Str.cpp
@@ -1031,14 +1031,14 @@ Vector<String> String::Split(const char* str, char separator, bool keepEmptyStri
         {
             const ptrdiff_t splitLen = splitEnd - str;
             if (splitLen > 0 || keepEmptyStrings)
-                ret.push_back(String(str, splitLen));
+                ret.push_back(String(str, (uint)splitLen));
             str = splitEnd + 1;
         }
     }
 
     const ptrdiff_t splitLen = strEnd - str;
     if (splitLen > 0 || keepEmptyStrings)
-        ret.push_back(String(str, splitLen));
+        ret.push_back(String(str, (uint)splitLen));
     
     return ret;
 }
